IGNITE-17555 Fix review 2
