+++ modules/configuration/src/main/java/org/apache/ignite/internal/configuration/ConfigurationChanger.java
@@ -199,7 +199,7 @@ public abstract class ConfigurationChanger implements DynamicConfigurationChange
         Data data;
 
         try {
-            data = storage.readAllOnStart().get();
+            data = storage.readDataOnRecovery().get();
         } catch (ExecutionException e) {
             throw new ConfigurationChangeException("Failed to initialize configuration: " + e.getCause().getMessage(), e.getCause());
         } catch (InterruptedException e) {
@@ -573,22 +573,23 @@ public abstract class ConfigurationChanger implements DynamicConfigurationChange
         storageRoots = new StorageRoots(newSuperRoot, newChangeId);
 
         // Save revisions for recovery.
-        return storage.writeConfigurationRevision(oldStorageRoots.version, storageRoots.version).thenCompose(unused -> {
-            if (dataValuesPrefixMap.isEmpty()) {
-                oldStorageRoots.changeFuture.complete(null);
+        return storage.writeConfigurationRevision(oldStorageRoots.version, storageRoots.version)
+                .thenCompose(unused -> {
+                    if (dataValuesPrefixMap.isEmpty()) {
+                        oldStorageRoots.changeFuture.complete(null);
 
-                return CompletableFuture.completedFuture(null);
-            } else {
-                return notificator.notify(oldSuperRoot, newSuperRoot, newChangeId, notificationListenerCnt.incrementAndGet())
-                        .whenComplete((v, t) -> {
+                        return CompletableFuture.completedFuture(null);
+                    } else {
+                        long notificationNumber = notificationListenerCnt.incrementAndGet();
+                        return notificator.notify(oldSuperRoot, newSuperRoot, newChangeId, notificationNumber).whenComplete((v, t) -> {
                             if (t == null) {
                                 oldStorageRoots.changeFuture.complete(null);
                             } else {
                                 oldStorageRoots.changeFuture.completeExceptionally(t);
                             }
                         });
-            }
-        });
+                    }
+                });
     }
 
     /**
